pyproject.toml
[project]
name = "fashion-lens"
version = "0.1.0"
description = "AI Garment Classification & Inspiration Web App"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "pydantic>=2.0.0",
    "langgraph>=0.0.26",
    "langchain-google-genai>=0.0.9",
    "langchain-core>=0.1.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
managed = true
